Move pathfinder to its own module to fix cyclic dependecies.
