+++ opengever/base/utils.py
@@ -1,14 +1,11 @@
 from Acquisition import aq_inner
 from Acquisition import aq_parent
-from App.config import DefaultConfiguration
-from opengever.dossier.behaviors.dossier import IDossierMarker
+from opengever.dossier.interfaces import IDossierMarker
 from plone import api
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.interfaces.siteroot import IPloneSiteRoot
 from xml.sax.saxutils import escape
 from zope.component import getMultiAdapter
-import App.config
-import os
 
 
 def language_cache_key(method, context, request):
@@ -78,40 +75,7 @@ def get_preferred_language_code():
     return language_code.split('-')[0]
 
 
-class PathFinder(object):
-    """Helper class to provide various Zope2 Instance related paths that are
-    otherwise cumbersone to access.
-    """
 
-    def __init__(self):
-        self._assert_proper_configuration()
-        self._instance_home = os.environ['INSTANCE_HOME']
-        self._client_home = os.environ['CLIENT_HOME']
-
-    def _assert_proper_configuration(self):
-        cfg = App.config._config
-        if cfg is None or isinstance(cfg, DefaultConfiguration):
-            raise RuntimeError(
-                "Zope is not configured properly yet, refusing "
-                "operate on paths that might be wrong!")
-
-    @property
-    def var(self):
-        """Path to {buildout}/var
-        """
-        return os.path.normpath(os.path.join(self._client_home, '..'))
-
-    @property
-    def var_log(self):
-        """Path to {buildout}/var/log
-        """
-        return os.path.join(self.var, 'log')
-
-    @property
-    def buildout(self):
-        """Path to {buildout}
-        """
-        return os.path.normpath(os.path.join(self.var, '..'))
 
 
 def get_hostname(request):
